requirements.txt
python-dotenv
azure-search-documents==11.6.0b1
azure-storage-blob
azure-identity
azure-mgmt-web
azure-ai-vision-imageanalysis
requests